Moved BoxWhisker back to charts
