+++ holoviews/element/chart.py
@@ -319,3 +319,20 @@ class Area(Curve):
                                     new_type=Area)
             baseline += y
         return stacked
+
+
+class BoxWhisker(Chart):
+    """
+    BoxWhisker represent data as a distributions highlighting the
+    median, mean and various percentiles. It may have a single value
+    dimension and any number of key dimensions declaring the grouping
+    of each violin.
+    """
+
+    group = param.String(default='BoxWhisker', constant=True)
+
+    kdims = param.List(default=[], bounds=(0,None))
+
+    vdims = param.List(default=[Dimension('y')], bounds=(1,1))
+
+    _auto_indexable_1d = False
